requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Database & Async
sqlalchemy>=2.0.0
asyncpg>=0.28.0
aiosqlite>=0.17.0
alembic>=1.11.0

# Workers & Queue
celery>=5.3.0
redis>=5.0.0

# Scientific & ML
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
xarray>=2023.0.0
zarr>=2.16.0
segyio>=1.9.0

# Deep Learning (CPU defaults, use extra-index-url for CUDA)
torch>=2.0.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0

# Observability
prometheus-client>=0.17.0
structlog>=23.1.0
